backend/requirements.txt
flask>=2.3.3
langchain>=0.1.1
langchain-google-genai>=0.0.5
langchain-community>=0.0.16
langchain-core>=0.1.19
faiss-cpu>=1.7.4
google-generativeai>=0.3.1
openai>=1.3.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
langchain_openai
flask_cors
moviepy
Flask-Bcrypt==1.0.1
psycopg2==2.9.10
Flask-SQLAlchemy==3.1.1
py-jwt==2.10.1
moviepy==2.1.2
webrtcvad-wheels==2.0.14
pydub==0.25.1
elevenlabs==1.55.0
